[skip ci] moved profiler to be user facing (metrics window)
