+++ DearPyGui/src/core/mvApp.cpp
@@ -26,34 +26,6 @@ namespace Marvel {
 	mvApp* mvApp::s_instance = nullptr;
 	std::atomic_bool mvApp::s_started = false;
 
-
-	// utility structure for realtime plot
-	struct ScrollingBuffer {
-		int MaxSize;
-		int Offset;
-		ImVector<ImVec2> Data;
-		ScrollingBuffer() {
-			MaxSize = 2000;
-			Offset = 0;
-			Data.reserve(MaxSize);
-		}
-		void AddPoint(float x, float y) {
-			if (Data.size() < MaxSize)
-				Data.push_back(ImVec2(x, y));
-			else {
-				Data[Offset] = ImVec2(x, y);
-				Offset = (Offset + 1) % MaxSize;
-			}
-		}
-		void Erase() {
-			if (!Data.empty())
-			{
-				Data.shrink(0);
-				Offset = 0;
-			}
-		}
-	};
-
 	mvApp* mvApp::GetApp()
 	{
 		mvLog::Init();
@@ -193,10 +165,7 @@ namespace Marvel {
 		mvEventBus::Publish(mvEVT_CATEGORY_APP, mvEVT_RENDER);
 		mvEventBus::Publish(mvEVT_CATEGORY_APP, mvEVT_END_FRAME);
 
-#if defined(MV_PROFILE) && defined(MV_DEBUG)
         //mvEventBus::ShowDebug();
-		postProfile();
-#endif // MV_PROFILE
 	}
 
 	bool mvApp::checkIfMainThread() const
@@ -210,41 +179,6 @@ namespace Marvel {
 			return false;
 		}
 		return true;
-	}
-
-	void mvApp::postProfile()
-	{
-		static std::map<std::string, ScrollingBuffer> buffers;
-		static float t = 0;
-		t += ImGui::GetIO().DeltaTime;
-
-		const auto& results = mvInstrumentor::Get().getResults();
-
-		for (const auto& item : results)
-			buffers[item.first].AddPoint(t, (float)item.second.count());
-
-		//ImGui::SetNextWindowFocus();
-		ImGui::SetNextWindowSize(ImVec2(500, 500), ImGuiCond_FirstUseEver);
-		if (!ImGui::Begin("Profiling", nullptr))
-		{
-			ImGui::End();
-			return;
-		}
-
-		static float history = 10.0f;
-		ImGui::SliderFloat("History", &history, 1, 30, "%.1f s");
-
-		static ImPlotAxisFlags rt_axis = ImPlotAxisFlags_NoTickLabels;
-		ImPlot::SetNextPlotLimitsX(t - history, t, ImGuiCond_Always);
-		if (ImPlot::BeginPlot("##Scrolling", nullptr, nullptr, ImVec2(-1, -1), 0, rt_axis, 0 | ImPlotAxisFlags_LockMin)) {
-
-			for (const auto& item : results)
-				ImPlot::PlotLine(item.first.c_str(), &buffers[item.first].Data[0].x, &buffers[item.first].Data[0].y, buffers[item.first].Data.size(), buffers[item.first].Offset, 2 * sizeof(float));
-			ImPlot::EndPlot();
-		}
-
-		ImGui::End();
-
 	}
 
 	std::map<std::string, mvPythonParser>& mvApp::getParsers()
